style: change product price from bigdecimal to double

+++ src/main/java/com/wesleybertipaglia/products/dtos/ProductRecordDto.java
@@ -1,9 +1,7 @@
 package com.wesleybertipaglia.products.dtos;
 
-import java.math.BigDecimal;
-
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
-public record ProductRecordDto(@NotBlank String name, @NotNull BigDecimal price) {
+public record ProductRecordDto(@NotBlank String name, @NotNull double price) {
 }
